backend/requirements.txt
fastapi
uvicorn
pymongo
transformers
torch
pydantic
pandas
python-dotenv
google-generativeai
